Refactor transpiler visitors into modules
