+++ backend/src/core/transpiler/js_nodes/atributo.py
@@ -1,0 +1,2 @@
+def visit_atributo(self, nodo):
+    self.agregar_linea(self.obtener_valor(nodo))
