+++ backend/src/core/transpiler/js_nodes/funcion.py
@@ -1,0 +1,14 @@
+def visit_funcion(self, nodo):
+    """Transpila una definición de función en JavaScript."""
+    parametros = ", ".join(nodo.parametros)
+    cuerpo = nodo.cuerpo
+    if self.usa_indentacion is None:
+        self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
+    self.agregar_linea(f"function {nodo.nombre}({parametros}) {{")
+    if self.usa_indentacion:
+        self.indentacion += 1
+    for instruccion in cuerpo:
+        instruccion.aceptar(self)
+    if self.usa_indentacion:
+        self.indentacion -= 1
+    self.agregar_linea("}")
