+++ backend/src/core/transpiler/js_nodes/identificador.py
@@ -1,0 +1,2 @@
+def visit_identificador(self, nodo):
+    self.agregar_linea(self.obtener_valor(nodo))
