+++ backend/src/core/transpiler/js_nodes/imprimir.py
@@ -1,0 +1,3 @@
+def visit_imprimir(self, nodo):
+    valor = self.obtener_valor(nodo.expresion)
+    self.agregar_linea(f"console.log({valor});")
