+++ backend/src/core/transpiler/js_nodes/instancia.py
@@ -1,0 +1,2 @@
+def visit_instancia(self, nodo):
+    self.agregar_linea(f"{self.obtener_valor(nodo)};")
