+++ backend/src/core/transpiler/js_nodes/operacion_binaria.py
@@ -1,0 +1,2 @@
+def visit_operacion_binaria(self, nodo):
+    self.agregar_linea(self.obtener_valor(nodo))
