+++ backend/src/core/transpiler/js_nodes/retorno.py
@@ -1,0 +1,3 @@
+def visit_retorno(self, nodo):
+    valor = self.obtener_valor(nodo.expresion)
+    self.agregar_linea(f"return {valor};")
