+++ backend/src/core/transpiler/js_nodes/throw.py
@@ -1,0 +1,3 @@
+def visit_throw(self, nodo):
+    valor = self.obtener_valor(nodo.expresion)
+    self.agregar_linea(f"throw {valor};")
