+++ backend/src/core/transpiler/js_nodes/valor.py
@@ -1,0 +1,2 @@
+def visit_valor(self, nodo):
+    self.agregar_linea(self.obtener_valor(nodo))
