+++ backend/src/core/transpiler/python_nodes/atributo.py
@@ -1,0 +1,2 @@
+def visit_atributo(self, nodo):
+    self.codigo += f"{self.obtener_valor(nodo)}\n"
