+++ backend/src/core/transpiler/python_nodes/bucle_mientras.py
@@ -1,0 +1,7 @@
+def visit_bucle_mientras(self, nodo):
+    condicion = self.obtener_valor(nodo.condicion)
+    self.codigo += f"{self.obtener_indentacion()}while {condicion}:\n"
+    self.nivel_indentacion += 1
+    for instruccion in nodo.cuerpo:
+        instruccion.aceptar(self)
+    self.nivel_indentacion -= 1
