+++ backend/src/core/transpiler/python_nodes/holobit.py
@@ -1,0 +1,6 @@
+def visit_holobit(self, nodo):
+    valores = ", ".join(self.obtener_valor(v) for v in nodo.valores)
+    if nodo.nombre:
+        self.codigo += f"{nodo.nombre} = holobit([{valores}])\n"
+    else:
+        self.codigo += f"holobit([{valores}])\n"
