+++ backend/src/core/transpiler/python_nodes/identificador.py
@@ -1,0 +1,2 @@
+def visit_identificador(self, nodo):
+    self.codigo += self.obtener_valor(nodo)
