+++ backend/src/core/transpiler/python_nodes/imprimir.py
@@ -1,0 +1,3 @@
+def visit_imprimir(self, nodo):
+    valor = self.obtener_valor(getattr(nodo, "expresion", nodo))
+    self.codigo += f"{self.obtener_indentacion()}print({valor})\n"
