+++ backend/src/core/transpiler/python_nodes/instancia.py
@@ -1,0 +1,2 @@
+def visit_instancia(self, nodo):
+    self.codigo += f"{self.obtener_valor(nodo)}\n"
