+++ backend/src/core/transpiler/python_nodes/lista.py
@@ -1,0 +1,5 @@
+def visit_lista(self, nodo):
+    elementos = ", ".join(
+        self.obtener_valor(elemento) for elemento in nodo.elementos
+    )
+    self.codigo += f"[{elementos}]\n"
