+++ backend/src/core/transpiler/python_nodes/llamada_funcion.py
@@ -1,0 +1,3 @@
+def visit_llamada_funcion(self, nodo):
+    argumentos = ", ".join(self.obtener_valor(arg) for arg in nodo.argumentos)
+    self.codigo += f"{nodo.nombre}({argumentos})\n"
