+++ backend/src/core/transpiler/python_nodes/operacion_binaria.py
@@ -1,0 +1,2 @@
+def visit_operacion_binaria(self, nodo):
+    self.codigo += self.obtener_valor(nodo)
