+++ backend/src/core/transpiler/python_nodes/throw.py
@@ -1,0 +1,3 @@
+def visit_throw(self, nodo):
+    valor = self.obtener_valor(nodo.expresion)
+    self.codigo += f"{self.obtener_indentacion()}raise Exception({valor})\n"
