+++ backend/src/core/transpiler/python_nodes/valor.py
@@ -1,0 +1,2 @@
+def visit_valor(self, nodo):
+    self.codigo += self.obtener_valor(nodo)
