+++ backend/src/core/transpiler/to_js.py
@@ -16,6 +16,33 @@ from src.core.parser import Parser
 from src.core.lexer import TipoToken, Lexer
 from src.core.visitor import NodeVisitor
 
+from .js_nodes.asignacion import visit_asignacion as _visit_asignacion
+from .js_nodes.condicional import visit_condicional as _visit_condicional
+from .js_nodes.bucle_mientras import visit_bucle_mientras as _visit_bucle_mientras
+from .js_nodes.funcion import visit_funcion as _visit_funcion
+from .js_nodes.llamada_funcion import visit_llamada_funcion as _visit_llamada_funcion
+from .js_nodes.hilo import visit_hilo as _visit_hilo
+from .js_nodes.llamada_metodo import visit_llamada_metodo as _visit_llamada_metodo
+from .js_nodes.imprimir import visit_imprimir as _visit_imprimir
+from .js_nodes.retorno import visit_retorno as _visit_retorno
+from .js_nodes.holobit import visit_holobit as _visit_holobit
+from .js_nodes.for_ import visit_for as _visit_for
+from .js_nodes.lista import visit_lista as _visit_lista
+from .js_nodes.diccionario import visit_diccionario as _visit_diccionario
+from .js_nodes.elemento import visit_elemento as _visit_elemento
+from .js_nodes.clase import visit_clase as _visit_clase
+from .js_nodes.metodo import visit_metodo as _visit_metodo
+from .js_nodes.try_catch import visit_try_catch as _visit_try_catch
+from .js_nodes.throw import visit_throw as _visit_throw
+from .js_nodes.importar import visit_import as _visit_import
+from .js_nodes.instancia import visit_instancia as _visit_instancia
+from .js_nodes.atributo import visit_atributo as _visit_atributo
+from .js_nodes.operacion_binaria import visit_operacion_binaria as _visit_operacion_binaria
+from .js_nodes.operacion_unaria import visit_operacion_unaria as _visit_operacion_unaria
+from .js_nodes.valor import visit_valor as _visit_valor
+from .js_nodes.identificador import visit_identificador as _visit_identificador
+from .js_nodes.para import visit_para as _visit_para
+
 
 class TranspiladorJavaScript(NodeVisitor):
     def __init__(self):
@@ -70,263 +97,34 @@ class TranspiladorJavaScript(NodeVisitor):
             nodo.aceptar(self)
         return "\n".join(self.codigo)
 
+
+# Asignar los visitantes externos a la clase
+TranspiladorJavaScript.visit_asignacion = _visit_asignacion
+TranspiladorJavaScript.visit_condicional = _visit_condicional
+TranspiladorJavaScript.visit_bucle_mientras = _visit_bucle_mientras
+TranspiladorJavaScript.visit_funcion = _visit_funcion
+TranspiladorJavaScript.visit_llamada_funcion = _visit_llamada_funcion
+TranspiladorJavaScript.visit_hilo = _visit_hilo
+TranspiladorJavaScript.visit_llamada_metodo = _visit_llamada_metodo
+TranspiladorJavaScript.visit_imprimir = _visit_imprimir
+TranspiladorJavaScript.visit_retorno = _visit_retorno
+TranspiladorJavaScript.visit_holobit = _visit_holobit
+TranspiladorJavaScript.visit_for = _visit_for
+TranspiladorJavaScript.visit_lista = _visit_lista
+TranspiladorJavaScript.visit_diccionario = _visit_diccionario
+TranspiladorJavaScript.visit_elemento = _visit_elemento
+TranspiladorJavaScript.visit_clase = _visit_clase
+TranspiladorJavaScript.visit_metodo = _visit_metodo
+TranspiladorJavaScript.visit_try_catch = _visit_try_catch
+TranspiladorJavaScript.visit_throw = _visit_throw
+TranspiladorJavaScript.visit_import = _visit_import
+TranspiladorJavaScript.visit_instancia = _visit_instancia
+TranspiladorJavaScript.visit_atributo = _visit_atributo
+TranspiladorJavaScript.visit_operacion_binaria = _visit_operacion_binaria
+TranspiladorJavaScript.visit_operacion_unaria = _visit_operacion_unaria
+TranspiladorJavaScript.visit_valor = _visit_valor
+TranspiladorJavaScript.visit_identificador = _visit_identificador
+TranspiladorJavaScript.visit_para = _visit_para
+
     # Métodos de transpilación para tipos de nodos básicos
 
-    def visit_asignacion(self, nodo):
-        """Transpila una asignación en JavaScript."""
-        if self.usa_indentacion is None:
-            self.usa_indentacion = hasattr(nodo, "variable") or hasattr(nodo, "identificador")
-        nombre_raw = getattr(nodo, "identificador", getattr(nodo, "variable", None))
-        if isinstance(nombre_raw, NodoAtributo):
-            nombre = self.obtener_valor(nombre_raw)
-            prefijo = ""
-        else:
-            nombre = nombre_raw
-            prefijo = "let " if hasattr(nodo, "variable") else ""
-        valor = getattr(nodo, "expresion", getattr(nodo, "valor", None))
-        valor = self.obtener_valor(valor)
-        self.agregar_linea(f"{prefijo}{nombre} = {valor};")
-
-    def visit_condicional(self, nodo):
-        """Transpila un 'if-else' con condiciones compuestas."""
-        cuerpo_si = getattr(nodo, "cuerpo_si", getattr(nodo, "bloque_si", []))
-        cuerpo_sino = getattr(nodo, "cuerpo_sino", getattr(nodo, "bloque_sino", []))
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(
-                hasattr(ins, "variable") for ins in cuerpo_si + cuerpo_sino
-            )
-        condicion = self.obtener_valor(nodo.condicion)
-        self.agregar_linea(f"if ({condicion}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo_si:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        if cuerpo_sino:
-            if self.usa_indentacion:
-                self.agregar_linea("} else {")
-            else:
-                self.agregar_linea("}")
-                self.agregar_linea("else {")
-            if self.usa_indentacion:
-                self.indentacion += 1
-            for instruccion in cuerpo_sino:
-                instruccion.aceptar(self)
-            if self.usa_indentacion:
-                self.indentacion -= 1
-            self.agregar_linea("}")
-        else:
-            self.agregar_linea("}")
-
-    def visit_bucle_mientras(self, nodo):
-        """Transpila un bucle 'while' en JavaScript, permitiendo anidación."""
-        cuerpo = nodo.cuerpo
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
-        condicion = self.obtener_valor(nodo.condicion)
-        self.agregar_linea(f"while ({condicion}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
-
-    def visit_funcion(self, nodo):
-        """Transpila una definición de función en JavaScript."""
-        parametros = ", ".join(nodo.parametros)
-        cuerpo = nodo.cuerpo
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
-        self.agregar_linea(f"function {nodo.nombre}({parametros}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
-
-    def visit_llamada_funcion(self, nodo):
-        """Transpila una llamada a función en JavaScript."""
-        parametros = ", ".join(self.obtener_valor(a) for a in nodo.argumentos)
-        self.agregar_linea(f"{nodo.nombre}({parametros});")
-
-    def visit_hilo(self, nodo):
-        args = ", ".join(self.obtener_valor(a) for a in nodo.llamada.argumentos)
-        self.agregar_linea(f"Promise.resolve().then(() => {nodo.llamada.nombre}({args}));")
-
-    def visit_llamada_metodo(self, nodo):
-        args = ", ".join(self.obtener_valor(a) for a in nodo.argumentos)
-        obj = self.obtener_valor(nodo.objeto)
-        self.agregar_linea(f"{obj}.{nodo.nombre_metodo}({args});")
-
-    def visit_imprimir(self, nodo):
-        valor = self.obtener_valor(nodo.expresion)
-        self.agregar_linea(f"console.log({valor});")
-
-    def visit_retorno(self, nodo):
-        valor = self.obtener_valor(nodo.expresion)
-        self.agregar_linea(f"return {valor};")
-
-    def visit_holobit(self, nodo):
-        """Transpila una asignación de Holobit en JavaScript."""
-        valores = ", ".join(map(str, nodo.valores))
-        nombre = getattr(nodo, "nombre", None)
-        if nombre:
-            self.agregar_linea(f"let {nombre} = new Holobit([{valores}]);")
-        else:
-            self.agregar_linea(f"new Holobit([{valores}]);")
-
-    # Métodos de transpilación para nuevas estructuras avanzadas
-
-    def visit_for(self, nodo):
-        """Transpila un bucle 'for...of' en JavaScript, permitiendo anidación."""
-        cuerpo = nodo.cuerpo
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
-        iterable = self.obtener_valor(nodo.iterable)
-        self.agregar_linea(f"for (let {nodo.variable} of {iterable}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
-
-    def visit_lista(self, nodo):
-        """Transpila una lista en JavaScript, permitiendo anidación."""
-        elementos = ", ".join(self.visit_elemento(e) for e in nodo.elementos)
-        self.agregar_linea(f"[{elementos}]")
-
-    def visit_diccionario(self, nodo):
-        """Transpila un diccionario en JavaScript, permitiendo anidación."""
-        elementos = getattr(nodo, "pares", getattr(nodo, "elementos", []))
-        pares = ", ".join([
-            f"{clave}: {self.visit_elemento(valor)}" for clave, valor in elementos
-        ])
-        self.agregar_linea(f"{{{pares}}}")
-
-    def visit_elemento(self, elemento):
-        """Transpila un elemento o estructura."""
-        if isinstance(elemento, NodoLista):
-            transpiled_element = []
-            self.codigo, original_codigo = transpiled_element, self.codigo
-            self.visit_lista(elemento)
-            self.codigo = original_codigo
-            return ''.join(transpiled_element)
-        elif isinstance(elemento, NodoDiccionario):
-            transpiled_element = []
-            self.codigo, original_codigo = transpiled_element, self.codigo
-            self.visit_diccionario(elemento)
-            self.codigo = original_codigo
-            return ''.join(transpiled_element)
-        else:
-            return str(elemento)
-
-    def visit_clase(self, nodo):
-        """Transpila una clase en JavaScript, admitiendo métodos y clases anidados."""
-        metodos = getattr(nodo, "cuerpo", getattr(nodo, "metodos", []))
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(m, "variable") for m in metodos)
-        self.agregar_linea(f"class {nodo.nombre} {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for metodo in metodos:
-            metodo.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
-
-    def visit_metodo(self, nodo):
-        """Transpila un método de clase, permitiendo anidación."""
-        parametros = ", ".join(nodo.parametros)
-        cuerpo = nodo.cuerpo
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
-        self.agregar_linea(f"{nodo.nombre}({parametros}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
-
-    def visit_try_catch(self, nodo):
-        self.agregar_linea("try {")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in nodo.bloque_try:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        if nodo.bloque_catch:
-            catch_var = nodo.nombre_excepcion or ""
-            if self.usa_indentacion:
-                self.agregar_linea(f"}} catch ({catch_var}) {{")
-            else:
-                self.agregar_linea("}")
-                self.agregar_linea(f"catch ({catch_var}) {{")
-            if self.usa_indentacion:
-                self.indentacion += 1
-            for instruccion in nodo.bloque_catch:
-                instruccion.aceptar(self)
-            if self.usa_indentacion:
-                self.indentacion -= 1
-            self.agregar_linea("}")
-        else:
-            self.agregar_linea("}")
-
-    def visit_throw(self, nodo):
-        valor = self.obtener_valor(nodo.expresion)
-        self.agregar_linea(f"throw {valor};")
-
-    def visit_import(self, nodo):
-        """Carga y transpila el módulo indicado."""
-        try:
-            with open(nodo.ruta, "r", encoding="utf-8") as f:
-                codigo = f.read()
-        except FileNotFoundError:
-            raise FileNotFoundError(f"Módulo no encontrado: {nodo.ruta}")
-
-        lexer = Lexer(codigo)
-        tokens = lexer.analizar_token()
-        ast = Parser(tokens).parsear()
-        for subnodo in ast:
-            subnodo.aceptar(self)
-
-    def visit_instancia(self, nodo):
-        self.agregar_linea(f"{self.obtener_valor(nodo)};")
-
-    def visit_atributo(self, nodo):
-        self.agregar_linea(self.obtener_valor(nodo))
-
-    def visit_operacion_binaria(self, nodo):
-        self.agregar_linea(self.obtener_valor(nodo))
-
-    def visit_operacion_unaria(self, nodo):
-        self.agregar_linea(self.obtener_valor(nodo))
-
-    def visit_valor(self, nodo):
-        self.agregar_linea(self.obtener_valor(nodo))
-
-    def visit_identificador(self, nodo):
-        self.agregar_linea(self.obtener_valor(nodo))
-
-    def visit_para(self, nodo):
-        cuerpo = nodo.cuerpo
-        if self.usa_indentacion is None:
-            self.usa_indentacion = any(hasattr(ins, "variable") for ins in cuerpo)
-        iterable = self.obtener_valor(nodo.iterable)
-        self.agregar_linea(f"for (let {nodo.variable} of {iterable}) {{")
-        if self.usa_indentacion:
-            self.indentacion += 1
-        for instruccion in cuerpo:
-            instruccion.aceptar(self)
-        if self.usa_indentacion:
-            self.indentacion -= 1
-        self.agregar_linea("}")
